+++ R/lifeduration.R
@@ -1,11 +1,11 @@
-#' Package life duration at specific Date or for a specific version
+#' Package version life duration at specific Date or for a specific version
 #' @description using CRAN website to get a package life duration for certain version or at a specific Date.
 #' @param pac character a package name.
 #' @param version character version of package. Default: NULL
 #' @param at Date old version of package. Default: NULL
 #' @param lib.loc character vector. Is omitted for non NULL version. Default: NULL
 #' @param repos character the base URL of the repositories to use. Default `https://cran.rstudio.com/`
-#' @return logical if package is healthy.
+#' @return list named of `difftime`, number of days package version was the newest one.
 #' @note Function will scrap two CRAN URLS. Works only with CRAN packages.
 #' Please as a courtesy to the R CRAN, don't overload their server by constantly using this function.
 #' Results are cached for 1 hour with `memoise` package, memory cache.
@@ -54,17 +54,20 @@ pac_lifeduration <- function(pac, version = NULL, at = NULL, lib.loc = NULL, rep
   }
 }
 
-#' Packages life duration at specific Date or for a specific version
+#' Packages versions life duration at specific Date or for a specific version
 #' @description using CRAN website to get packages life duration for versions or specific Date.
 #' @param pacs character vector packages names.
 #' @param versions character vector versions of packages. Default: NULL
 #' @param at Date old version of package. Default: NULL
 #' @param lib.loc character vector. Is omitted for non NULL version. Default: NULL
 #' @param repos character the base URL of the repositories to use. Default `https://cran.rstudio.com/`
-#' @return logical if package is healthy.
+#' @return list named of `difftime`, number of days package version was the newest one.
 #' @note Function will scrap two CRAN URLS. Works only with CRAN packages.
 #' Please as a courtesy to the R CRAN, don't overload their server by constantly using this function.
 #' Results are cached for 1 hour with `memoise` package, memory cache.
+#' There is used `parallel::mclapply` function.
+#' Remember that `mclaply` under Windows works like the regular `lapply` function.
+#' To set higher number of cores use code like `options(mc.cores = parallel::detectCores() - 2)` at the beginning of the session.
 #' @export
 #' @examples
 #' pacs_lifeduration(c("memoise", "rlang"))
@@ -74,7 +77,7 @@ pacs_lifeduration <- function(pacs, versions = NULL, at = NULL, lib.loc = NULL, 
   stopifnot(is.null(versions) || length(pacs) == length(versions))
 
   stats::setNames(
-    lapply(
+    parallel::mclapply(
       seq_along(pacs),
       function(x) pac_lifeduration(pacs[x], version = versions[x], at = at, lib.loc = lib.loc, repos = repos)
     ),
@@ -84,7 +87,7 @@ pacs_lifeduration <- function(pacs, versions = NULL, at = NULL, lib.loc = NULL, 
 
 #' Package health state at a specific Date or for a specific version
 #' @description using CRAN website to get a package version/versions used at a specific Date interval.
-#' A healthy package is a if it was published for more than x days where default is 14 days.
+#' A healthy package was published for more than x days, where default is 14 days.
 #' CRAN team gives around one/two week to resolved a package which gave errors under the check page.
 #' @param pac character a package name.
 #' @param version character version of package. Default: NULL
@@ -137,7 +140,7 @@ pac_health <- function(pac, version = NULL, at = NULL, limit = 14, lib.loc = NUL
 
 #' Packages health state at a specific Date or for a specific versions
 #' @description using CRAN website to get a package version/versions used at a specific Date interval.
-#' A healthy package is a if it was published for more than 14 days.
+#' A healthy package was published for more than x days, where default is 14 days.
 #' CRAN team gives around one week to resolved a package which gave errors under the check page.
 #' @param pacs character vector packages names.
 #' @param versions character vector versions of packages. Default: NULL
@@ -145,10 +148,13 @@ pac_health <- function(pac, version = NULL, at = NULL, limit = 14, lib.loc = NUL
 #' @param limit numeric at least days to treat as healthy. Default: 14
 #' @param lib.loc character vector. Is omitted for non NULL version. Default: NULL
 #' @param repos character the base URL of the repositories to use. Default `https://cran.rstudio.com/`
-#' @return logical vector, TRUE if a package is healthy.
+#' @return list named, TRUE if a package was published for more than x days.
 #' @note Function will scrap two CRAN URLS. Works only with CRAN packages.
 #' Please as a courtesy to the R CRAN, don't overload their server by constantly using this function.
 #' Results are cached for 1 hour with `memoise` package, memory cache.
+#' There is used `parallel::mclapply` function.
+#' Remember that `mclaply` under Windows works like the regular `lapply` function.
+#' To set higher number of cores use code like `options(mc.cores = parallel::detectCores() - 2)` at the beginning of the session.
 #' @export
 #' @examples
 #' pacs_health(c("memoise"))
@@ -158,7 +164,7 @@ pacs_health <- function(pacs, versions = NULL, at = NULL, limit = 14, lib.loc = 
   stopifnot(is.null(versions) || length(pacs) == length(versions))
 
   stats::setNames(
-    lapply(
+    parallel::mclapply(
       seq_along(pacs),
       function(x) pac_health(pacs[x], version = versions[x], at = at, limit = limit, lib.loc = lib.loc, repos = repos)
     ),